streamer/requirements.txt
pyzmq
docker